Move out some utility functions from the engine into their own utilities submodule
